+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+= 1.10.0, 2022-06-29 =
+* [subscr] Do not process subscription if there is no renewal date.
+* [embedded] Bug fix: Do not refresh payment page automatically after an unrecoverable error.
+* [franfinance] Send information about shipping method for Franfinance payment method.
+* Added shipping options configuration field.
+* Support for "woocommerce blocks" in standard redirection payment.
+* Bug fix: Fix payment methods display based on amount restrictions for commands created from WooCommerce Back Office.
+* Possibility to make refunds for payments.
+
 = 1.9.5, 2022-05-05 =
 * Update list of supported payment means.